requirements.txt
requests
pyyaml
beautifulsoup4
lxml
trafilatura
tqdm
sentence-transformers
faiss-cpu
numpy
rich
